Fixed bugs
A timer of really long periods would overflow before it had a chance to
set the registers. The registers would be set correctly, just with
incorrect values. This also meant removing the rounding I had as it
would also cause overflows.

Changed semanitic versioning warning into errors.

There was a bug in timer1 where it wasn't dividing by an actual 16 bit
timer, but rather an 8 bit timer.

There was a bug in the timer interrpts, where they would be called
several times in succession because because the interrupt flag was not
being set.